Implement FetchContests request.
