+++ app/src/main/java/com/bogdan/codeforceswatcher/feature/contests/ContestsFragment.kt
@@ -4,25 +4,19 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
-import com.bogdan.codeforceswatcher.CwApp
 import com.bogdan.codeforceswatcher.R
 import com.bogdan.codeforceswatcher.adapter.ContestAdapter
-import com.bogdan.codeforceswatcher.feature.contests.redux.ContestsActions
+import com.bogdan.codeforceswatcher.feature.contests.redux.ContestsRequests
 import com.bogdan.codeforceswatcher.feature.contests.redux.ContestsState
 import com.bogdan.codeforceswatcher.model.Contest
-import com.bogdan.codeforceswatcher.network.RestClient
 import com.bogdan.codeforceswatcher.store
 import com.bogdan.codeforceswatcher.util.Analytics
 import kotlinx.android.synthetic.main.fragment_contests.recyclerView
 import kotlinx.android.synthetic.main.fragment_contests.swipeToRefresh
 import org.rekotlin.StoreSubscriber
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 
 class ContestsFragment : Fragment(), SwipeRefreshLayout.OnRefreshListener,
     StoreSubscriber<ContestsState> {
@@ -44,7 +38,7 @@ class ContestsFragment : Fragment(), SwipeRefreshLayout.OnRefreshListener,
     }
 
     override fun onRefresh() {
-        updateContestList()
+        store.dispatch(ContestsRequests.FetchContests())
         Analytics.logContestsListRefresh()
     }
 
@@ -61,7 +55,7 @@ class ContestsFragment : Fragment(), SwipeRefreshLayout.OnRefreshListener,
 
     private fun initViews() {
         swipeToRefresh.setOnRefreshListener(this)
-        updateContestList(false)
+        store.dispatch(ContestsRequests.FetchContests())
 
         contestAdapter = ContestAdapter(listOf(), requireContext())
         recyclerView.adapter = contestAdapter
@@ -71,35 +65,5 @@ class ContestsFragment : Fragment(), SwipeRefreshLayout.OnRefreshListener,
         liveData.observe(this, Observer<List<Contest>> { contestList ->
             contestList?.let { contestsList -> contestAdapter.setItems(contestsList.sortedBy(Contest::time)) }
         })*/
-    }
-
-    private fun updateContestList(shouldDisplayError: Boolean = true) {
-        RestClient.getContests().enqueue(object : Callback<ContestResponse> {
-
-            override fun onResponse(
-                call: Call<ContestResponse>,
-                response: Response<ContestResponse>
-            ) {
-                if (response.body() != null) {
-                    val contestList = response.body()?.result
-                    if (contestList != null) {
-                        store.dispatch(ContestsActions.ContestsLoaded(contestList))
-                        // DatabaseClient.contestDao.deleteAll(contestList)
-                        // DatabaseClient.contestDao.insert(contestList)
-                    }
-                }
-                if (activity != null) swipeToRefresh.isRefreshing = false
-            }
-
-            override fun onFailure(call: Call<ContestResponse>, t: Throwable) {
-                if (activity != null) swipeToRefresh.isRefreshing = false
-                if (shouldDisplayError)
-                    Toast.makeText(
-                        CwApp.app,
-                        getString(R.string.no_internet_connection),
-                        Toast.LENGTH_SHORT
-                    ).show()
-            }
-        })
     }
 }
