--- app/src/main/java/com/bogdan/codeforceswatcher/feature/contests/redux/ContestsActions.kt
@@ -1,11 +1,0 @@
-package com.bogdan.codeforceswatcher.feature.contests.redux
-
-import com.bogdan.codeforceswatcher.model.Contest
-import org.rekotlin.Action
-
-class ContestsActions {
-
-    data class ContestsLoaded(
-        val contests: List<Contest>
-    ) : Action
-}
